fix import issues and polish up for publishing
